feat: make qty inputs integer-only across every module
Qty fields (never physical_qty/minimum_stock-style thresholds, just
transactional quantities) no longer accept decimals anywhere they can
be entered, front and back:

- PKB (work order) service/sparepart lines
- Invoice service/sparepart lines (both work-order-based and direct
  sale) and their edit forms
- Goods receipt lines, including the xlsx import path
- Stock adjustment physical_qty (min stays 0 — a real physical count
  of zero is valid, just no fractional count)
- Stock transfer lines

Two enforcement layers per field:
1. HTML: step="1" min="1" (min="0" for physical_qty) on every
   <input type="number"> qty field, replacing step="0.001" — the
   browser's native constraint validation now rejects "2.5" outright.
2. Backend: every FormRequest's qty rule changed from
   ['numeric', 'min:0.001'] to ['integer', 'min:1'] (or 'min:0' for
   physical_qty), so a decimal sent directly via API/curl is rejected
   with a 422 regardless of what the browser allowed. The goods-receipt
   xlsx importer (which builds its own validation instead of using a
   FormRequest) got the equivalent explicit check added, since
   is_numeric() alone doesn't reject "2.5".

Verified in-browser: a service-qty input now has step="1" min="1", and
typing 2.5 fails checkValidity() with the browser's native "nearest
valid values are 2 and 3" message. Added 8 tests (one/two per module)
proving the backend 422s a decimal qty sent directly, bypassing the
browser entirely. Full suite: 1071 passed, 0 failed.

+++ app/Http/Requests/StoreStockAdjustmentRequest.php
@@ -33,7 +33,7 @@ class StoreStockAdjustmentRequest extends FormRequest
             'lines' => ['required', 'array', 'min:1'],
             'lines.*' => ['array'],
             'lines.*.sparepart_branch_id' => ['required', 'integer', 'exists:sparepart_branches,id', 'distinct'],
-            'lines.*.physical_qty' => ['required', 'numeric', 'min:0'],
+            'lines.*.physical_qty' => ['required', 'integer', 'min:0'],
             'lines.*.reason' => ['required', 'string', 'max:255'],
         ];
     }
